ovsdb: refactor client into multiple files
